columnar: dissolve the `arrangement` submodule (Phase A)
`columnar/arrangement` held a confusing mix: substrate (`Coltainer`,
`trie_merger`), container-glue (`TrieChunker`), and trace-name aliases
(`Val*`) that re-exported `chunk::col` types — so naming a columnar trace
meant reaching "around" into `arrangement`, even though the trace lives in
`trace/chunk/col`. Dissolve it, sorting each piece to where it belongs:

- `Coltainer` → `columnar/layout.rs` (it's the `BatchContainer` half of
  `ColumnarLayout`; pure substrate).
- `trie_merger` → `columnar/trie_merger.rs` (the trie-native merge core, beside
  the `updates` trie it operates on).
- `TrieChunker` → `columnar/chunker.rs` (the `RecordedUpdates → ColChunk`
  container-glue chunker).
- The `Val*` aliases move to `columnar/mod.rs` as thin re-exports of the
  canonical `trace::chunk::col` harness types.

`chunk/col.rs` now imports substrate directly (`columnar::{layout, trie_merger,
updates, spill}`) with no lateral hop through `arrangement`. Pure motion: no
behavior change, build/tests clean, spill smoke unchanged.
